# public/openapi.yaml
openapi: 3.1.0
info:
  title: Aegis API
  version: 1.0.0
  summary: HTTP surface for the Aegis D2A social agent platform.
  description: |
    Aegis is a decentralized social agent platform that scores and curates
    content across Nostr, Farcaster, RSS, X, and arbitrary URLs using the
    V/C/L rubric (V=signal, C=context, L=slop) with a multi-tier scoring
    cascade (local LLMs -> BYOK Claude -> IC LLM -> server Claude ->
    heuristic fallback). This document describes the public HTTP endpoints
    exposed by the Next.js app at https://aegis-ai.xyz/api.
    D2A endpoints additionally implement the x402 payment protocol and
    CORS for cross-origin agent-to-agent consumption.
  license:
    name: MIT
    url: https://github.com/dwebxr/aegis/blob/main/LICENSE
  contact:
    name: Aegis on GitHub
    url: https://github.com/dwebxr/aegis
servers:
  - url: https://aegis-ai.xyz
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/d2a/briefing:
    get:
      tags: [D2A]
      summary: Get curated briefing (x402-paid)
      description: |
        Returns either a principal-specific briefing (`?principal=`) or a
        paginated global summary. When `X402_RECEIVER` is configured, the
        endpoint is wrapped in x402 v2 payment middleware and callers must
        include a `PAYMENT-SIGNATURE` header (see https://x402.org).
        If `X402_FREE_TIER_ENABLED` is `true`, `?preview=true` returns a
        truncated payload without payment. Supports CORS for allowlisted
        origins.
      operationId: getD2ABriefing
      security:
        - {}
        - x402: []
      parameters:
        - name: principal
          in: query
          required: false
          description: IC principal (text form) for a user-specific briefing.
          schema:
            type: string
        - name: preview
          in: query
          required: false
          description: When `X402_FREE_TIER_ENABLED`, set `true` to bypass payment.
          schema:
            type: string
            enum: ["true"]
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          required: false
          description: Per-user path default 50 / max 100; global path default 5 / max 10.
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: since
          in: query
          required: false
          description: ISO 8601 timestamp; excludes briefings generated before it.
          schema:
            type: string
            format: date-time
        - name: topics
          in: query
          required: false
          description: Comma-separated topic filter (case-insensitive, OR logic).
          schema:
            type: string
      responses:
        '200':
          description: Per-user or global briefing.
          headers:
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/D2ABriefingResponse'
                  - $ref: '#/components/schemas/GlobalBriefingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: >
            x402 payment required, or settlement failed after a successful
            handler response. Protocol details are carried in the payment headers.
          headers:
            PAYMENT-REQUIRED:
              $ref: '#/components/headers/PaymentRequired'
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/X402SettlementFailureResponse'
        '404':
          description: No briefing available for principal, or no global briefings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          description: Shared D2A payments are disabled by the operator kill switch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    options:
      tags: [D2A]
      summary: CORS preflight
      operationId: d2aBriefingOptions
      responses:
        '204':
          description: Preflight response with CORS headers.
...
  /api/d2a/briefing/changes:
    get:
      tags: [D2A]
      summary: Diff endpoint — item hashes since a timestamp
      description: |
        Lightweight diff returning `added` item hashes from briefings
        generated after `since`. Useful for agents polling for new content
        without fetching full payloads. Only returns `"action": "added"`
        (true diff with updated/removed requires server-side snapshot
        storage which is not implemented). When `X402_RECEIVER` is
        configured, the endpoint is wrapped in the same x402 payment
        middleware as `/api/d2a/briefing`, using an x402 v2
        `PAYMENT-SIGNATURE` header. If `X402_FREE_TIER_ENABLED` is
        `true`, `?preview=true` returns the diff without payment but with
        `title`/`sourceUrl` redacted (hashes and ranks only).
      operationId: getD2ABriefingChanges
      security:
        - {}
        - x402: []
      parameters:
        - name: since
          in: query
          required: true
          description: ISO 8601 timestamp.
          schema:
            type: string
            format: date-time
        - name: preview
          in: query
          required: false
          description: >
            When `X402_FREE_TIER_ENABLED`, set `true` to bypass payment;
            `title`/`sourceUrl` are redacted to empty strings.
          schema:
            type: string
            enum: ["true"]
      responses:
        '200':
          description: List of new items since `since`.
          headers:
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: >
            x402 payment required, or settlement failed after a successful
            handler response. Protocol details are carried in the payment headers.
          headers:
            PAYMENT-REQUIRED:
              $ref: '#/components/headers/PaymentRequired'
            PAYMENT-RESPONSE:
              $ref: '#/components/headers/PaymentResponse'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/X402SettlementFailureResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          description: Shared D2A payments are disabled by the operator kill switch.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    options:
      tags: [D2A]
      summary: CORS preflight
      operationId: d2aBriefingChangesOptions
      responses:
        '204':
          description: Preflight response with CORS headers.
...
components:
  securitySchemes:
    byokApiKey:
      type: apiKey
      in: header
      name: x-user-api-key
      description: |
        User's Anthropic API key ("Bring Your Own Key"). Must start with
        `sk-ant-`. When present, server budget and operator key are
        bypassed.
    pushToken:
      type: apiKey
      in: header
      name: token
      description: |
        HMAC-SHA256 push token obtained from `/api/push/token`. (Note: the
        token is passed in the JSON body, not a header, on /api/push/send;
        this security scheme documents the credential semantically.)
    x402:
      type: apiKey
      in: header
      name: PAYMENT-SIGNATURE
      description: |
        Base64 x402 v2 payment payload (see https://x402.org). Applies to
        `/api/d2a/briefing`, `/api/d2a/briefing/changes`, and
        `/api/d2a/score` when their payment gates are enabled. Briefing
        preview routes can bypass payment when their free tier is enabled;
        score requires payment in production.
    openPayX402:
      type: apiKey
      in: header
      name: X-PAYMENT
      description: >
        OpenPay-flavored x402 v1 payment payload for
        `/api/d2a/briefing-jpyc` on Polygon.
  headers:
    PaymentRequired:
      description: Base64-encoded x402 v2 PaymentRequired object.
      schema:
        type: string
    PaymentResponse:
      description: Base64-encoded x402 v2 settlement result.
      schema:
        type: string
  responses:
    BadRequest:
      description: Invalid request (missing/malformed input).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Too many requests — see `Retry-After` header.
      headers:
        Retry-After:
          schema:
            type: integer
            description: Seconds until the rate-limit window resets.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        reason:
          type: string
    X402SettlementFailureResponse:
      type: object
      description: >
        x402 v2 error body. The library currently emits an empty JSON object
        for unpaid and settlement-failure responses; PAYMENT-REQUIRED or
        PAYMENT-RESPONSE is authoritative.
      additionalProperties: true
...
    Verdict:
      type: string
      enum: [quality, slop]
...
    ScoreBreakdown:
      type: object
      required: [originality, insight, credibility, composite]
      properties:
        originality:
          type: number
          minimum: 0
          maximum: 10
        insight:
          type: number
          minimum: 0
          maximum: 10
        credibility:
          type: number
          minimum: 0
          maximum: 10
        composite:
          type: number
          minimum: 0
          maximum: 10
...
    D2ABriefingItem:
      type: object
      required: [title, content, source, sourceUrl, scores, verdict, reason, topics, briefingScore]
      properties:
        title:
          type: string
          description: >
            First ~80 characters of the item text (items have no standalone
            title); a trailing `…` marks truncation. The full text is in
            `content`.
        content:
          type: string
          description: Full item text (truncated only under `preview=true`).
        source:
          type: string
        sourceUrl:
          type: string
          format: uri
        scores:
          allOf:
            - $ref: '#/components/schemas/ScoreBreakdown'
            - type: object
              properties:
                vSignal:
                  type: number
                cContext:
                  type: number
                lSlop:
                  type: number
        verdict:
          $ref: '#/components/schemas/Verdict'
        reason:
          type: string
        topics:
          type: array
          items:
            type: string
        briefingScore:
          type: number
          description: >
            Briefing-relative rank score, normalized server-side at read time:
            1 = top priority item, others proportional in [0.0001, 1] (ranked
            items are always strictly positive — deeply-decayed ones floor at
            0.0001 and may tie there). Only 0 for a malformed/unscored entry.
            (The raw on-chain value carries a publish-time recency-decay factor
            that collapses toward 0 for old content — only its ORDER is
            meaningful, hence the normalization.) `serendipityPick.briefingScore`
            is a serendipity score on a different (0–10-ish) scale and is not
            normalized.
    D2ABriefingResponse:
      type: object
      required: [version, generatedAt, source, sourceUrl, summary, items, serendipityPick, meta]
      properties:
        version:
          type: string
          enum: ["1.0"]
        generatedAt:
          type: string
          format: date-time
        source:
          type: string
          enum: [aegis]
        sourceUrl:
          type: string
          format: uri
        summary:
          type: object
          required: [totalEvaluated, totalBurned, qualityRate]
          properties:
            totalEvaluated:
              type: integer
            totalBurned:
              type: integer
            qualityRate:
              type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/D2ABriefingItem'
        serendipityPick:
          oneOf:
            - $ref: '#/components/schemas/D2ABriefingItem'
            - type: 'null'
        meta:
          type: object
          required: [scoringModel, nostrPubkey, topics]
          properties:
            scoringModel:
              type: string
            nostrPubkey:
              type: string
              nullable: true
            topics:
              type: array
              items:
                type: string
    BriefingChange:
      type: object
      required: [action, itemHash, title, sourceUrl, composite, generatedAt]
      properties:
        action:
          type: string
          enum: [added]
        itemHash:
          type: string
        title:
          type: string
          description: Empty string when `preview=true` redaction applies.
        sourceUrl:
          type: string
          description: >
            Item URL. Empty string when `preview=true` redaction applies,
            so no `format: uri` constraint.
        composite:
          type: number
        generatedAt:
          type: string
          format: date-time
    ChangesResponse:
      type: object
      required: [since, checkedAt, changes]
      properties:
        since:
          type: string
          format: date-time
        checkedAt:
          type: string
          format: date-time
        changes:
          type: array
          items:
            $ref: '#/components/schemas/BriefingChange'
    GlobalBriefingContributor:
      type: object
      required: [principal, generatedAt, summary, topItems]
      properties:
        principal:
          type: string
        generatedAt:
          type: string
          format: date-time
        summary:
          type: object
          required: [totalEvaluated, totalBurned, qualityRate]
          properties:
            totalEvaluated:
              type: integer
            totalBurned:
              type: integer
            qualityRate:
              type: number
        topItems:
          type: array
          description: >
            Ranked INDEX (top 3) with article URLs. For full items —
            `content`, `reason`, full scores — request this contributor's
            briefing directly with `?principal=<contributor principal>`.
          items:
            type: object
            required: [title, sourceUrl, topics, briefingScore, verdict]
            properties:
              title:
                type: string
                description: First ~80 chars of the item text; `…` marks truncation.
              sourceUrl:
                type: string
                description: >
                  Article URL. Empty string when the item has no URL or
                  under `preview=true` redaction (URLs are paid content,
                  matching /changes semantics).
              topics:
                type: array
                items:
                  type: string
              briefingScore:
                type: number
                description: Briefing-relative rank score (1 = contributor's top item).
              verdict:
                $ref: '#/components/schemas/Verdict'
    GlobalBriefingResponse:
      type: object
      required: [version, type, generatedAt, pagination, contributors, aggregatedTopics, totalEvaluated, totalQualityRate]
      properties:
        version:
          type: string
          enum: ["1.0"]
        type:
          type: string
          enum: [global]
        generatedAt:
          type: string
          format: date-time
        pagination:
          type: object
          required: [offset, limit, total, hasMore]
          properties:
            offset:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            hasMore:
              type: boolean
        contributors:
          type: array
          items:
            $ref: '#/components/schemas/GlobalBriefingContributor'
        aggregatedTopics:
          type: array
          items:
            type: string
        totalEvaluated:
          type: integer
        totalQualityRate:
          type: number
...